+++ app/routes/components/Textarea/demos/Demo01.tsx
@@ -1,5 +1,5 @@
 import { Textarea } from "@/components/Textarea";
 
 export default function Demo() {
-  return <Textarea />;
+  return <Textarea placeholder="Enter your message..." />;
 }