+++ app/routes/components/Textarea/index.tsx
@@ -5,8 +5,7 @@ import { DocLayout } from "@/layouts/DocLayout";
 import type { Route } from "./+types/index";
 
 import { changelog } from "./changelog.codegen";
-import Demo01 from "./demos/Demo01";
-import demo01SourceCode from "./demos/Demo01.source_code.codegen";
+import Doc from "./doc.mdx";
 import { sourceCode } from "./source_code.codegen";
 
 export function meta(_: Route.MetaArgs) {
@@ -36,11 +35,11 @@ export default function TextareaPage() {
       />
       {tabsValue === "doc" && (
         <DocLayout.Content>
-          <DocLayout.Live node={<Demo01 />} code={demo01SourceCode} />
+          <Doc />
         </DocLayout.Content>
       )}
       {tabsValue === "source" && (
-        <DocLayout.Content>
+        <DocLayout.Content variant="files">
           <DocLayout.Files files={sourceCode} />
         </DocLayout.Content>
       )}