+++ src/components/Textarea/index.ts
@@ -1,3 +1,3 @@
-// Textarea from copy-ui @ 2025-10-27.
+// Textarea from copy-ui @ 2025-11-28.
 
 export { Textarea } from "./textarea";